refactor(serve-inquirer): split the menu flows into focused modules
serveInquirer.ts owned ~1500 lines of interactive flows, which made it
hard to test, review and change safely. Move each cohesive flow into its
own module under src/menus/:

  - mainMenu       banner, option list and top-level dispatch
  - runners        tests, scripts, forge tests, flatten, function list
  - network        chain selector, env builder, account balance
  - settings       chain activation, RPC/keys, file exclusions
  - moreSettings   workflows, Foundry setup, plugin/custom-command entries
  - plugins        Hardhat plugin install/uninstall
  - mockContracts  mock creator plus its CLI flag format/parse/run helpers
  - deploymentScripts, verifyContract, customCommands
  - fileSelector   the shared recursive directory browser
  - types          the inquirer answer shapes the prompts are typed against

serveInquirer.ts keeps only the ICliArgs shape and the flag dispatcher,
and re-exports formatAddCustomMockContractFlag /
parseAddCustomMockContractFlag so its import surface is unchanged.
No CLI behaviour changes.
